Moved export modules out of main page
